+++ packages/client/src/network/utils/index.ts
@@ -1,7 +1,13 @@
+export {
+  didCancel as didActionCancel,
+  didComplete as didActionComplete,
+  didFail as didActionFail,
+  resolveState as resolveActionState,
+  waitForCompletion as waitForActionCompletion,
+} from './actions';
 export { useStream } from './hooks';
 export {
   waitForComponentValue,
   waitForComponentValueIn,
   waitForComponentValueUpdate,
 } from './utils';
-export { checkActionState, waitForActionCompletion } from './waitForActionCompletion';
